refactor: 관리자 메인페이지 수정 - 페이지 헤더 제거 - Router element 분리
